refactor(downloads): to ts
